The autoParagraphEmptyRoots extracted to util function.
