+++ packages/ckeditor5-engine/src/controller/datacontroller.js
@@ -24,6 +24,7 @@ import ViewDocument from '../view/document';
 import ViewDowncastWriter from '../view/downcastwriter';
 
 import ModelRange from '../model/range';
+import { autoParagraphEmptyRoots } from '../model/utils/autoparagraphing';
 
 /**
  * Controller for the data pipeline. The data pipeline controls how data is retrieved from the document
@@ -144,7 +145,7 @@ export default class DataController {
 		// Fix empty roots after DataController is 'ready' (note that init method could be decorated and stopped).
 		// We need to handle this event because initial data could be empty and post-fixer would not get triggered.
 		this.on( 'ready', () => {
-			this.model.enqueueChange( 'transparent', writer => this.model._autoparagraphEmptyRoots( writer ) );
+			this.model.enqueueChange( 'transparent', writer => autoParagraphEmptyRoots( writer ) );
 		}, { priority: 'lowest' } );
 	}
